# backend/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "ApproveJoinRequestOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "magic_link_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Magic Link Token"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "display_name",
          "phone_number",
          "status"
        ],
        "title": "ApproveJoinRequestOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "JoinRequestAnswerOut": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "label",
          "answer"
        ],
        "title": "JoinRequestAnswerOut",
        "type": "object"
      },
...
      "JoinRequestOut": {
        "properties": {
          "answers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/JoinRequestAnswerOut"
            },
            "title": "Answers",
            "type": "array"
          },
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By Name"
          },
          "attached_user_official_rsvp_count": {
            "default": 0,
            "title": "Attached User Official Rsvp Count",
            "type": "integer"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "onboarded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarded At"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "previously_archived": {
            "default": false,
            "title": "Previously Archived",
            "type": "boolean"
          },
          "rejected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected At"
          },
          "rejected_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected By Name"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "display_name",
          "phone_number",
          "submitted_at",
          "status"
        ],
        "title": "JoinRequestOut",
        "type": "object"
      },
      "JoinRequestStatusIn": {
        "properties": {
          "status": {
            "maxLength": 20,
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "JoinRequestStatusIn",
        "type": "object"
      },
...
      "NotificationOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_read": {
            "title": "Is Read",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "notification_type": {
            "title": "Notification Type",
            "type": "string"
          },
          "related_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related User Id"
          }
        },
        "required": [
          "id",
          "notification_type",
          "event_id",
          "related_user_id",
          "message",
          "is_read",
          "created_at"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
...
      "UnreadCountOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "UnreadCountOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "GatedJWTAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "OptionalJWTAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "",
    "title": "PDA API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/community/join-requests/": {
      "get": {
        "operationId": "community__join_requests_list_join_requests",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JoinRequestOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "List Join Requests",
        "tags": [
          "community"
        ]
      }
    },
    "/api/community/join-requests/{id}/": {
      "patch": {
        "operationId": "community__join_requests_update_join_request_status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRequestStatusIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveJoinRequestOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "Update Join Request Status",
        "tags": [
          "community"
        ]
      }
    },
...
    "/api/community/join-requests/{id}/unreject/": {
      "patch": {
        "operationId": "community__join_requests_unreject_join_request",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "Unreject Join Request",
        "tags": [
          "community"
        ]
      }
    },
...
    "/api/notifications/": {
      "get": {
        "operationId": "notifications_api_list_notifications",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/read-all/": {
      "post": {
        "operationId": "notifications_api_mark_all_read",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response",
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "Mark All Read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/unread-count/": {
      "get": {
        "operationId": "notifications_api_unread_count",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "Unread Count",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/{notification_id}/read/": {
      "post": {
        "operationId": "notifications_api_mark_read",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response",
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "GatedJWTAuth": []
          }
        ],
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    }
  },
  "servers": []
}